Improve code consistency and reduce coupling
## UV Activation Logic Clarification
- Add detailed documentation explaining intentional use of original blue value (b)
- Clarify why finalR/finalG are used but original b is checked for UV activation
- Prevents UV from being disabled when blue contributes to white channel
- Better explains the sophisticated color extraction logic

## Modal Height Decoupling
- Convert hardcoded MODAL_RESERVED_HEIGHT to optional maxHeight prop
- RoscoluxSwatchPicker no longer tightly coupled to ColorPickerModal layout
- Parent component can control height based on its own layout needs
- Maintains backwards compatibility with default value

## Eliminate Code Duplication
- Create shared colorHelpers.ts utility file
- Extract rgbToHex and hexToRgb functions to single source of truth
- Remove duplicate implementations from ColorWheelPicker and RoscoluxSwatchPicker
- Ensures consistent color conversion behavior across components

## Benefits
- More flexible component architecture with reduced coupling
- Single source of truth for color conversion utilities
- Clear documentation of intentional design decisions
- Easier maintenance and testing of shared functionality
